finished series

+++ laravel-57-from-scratch/app/Http/Controllers/ProjectsController.php
@@ -2,16 +2,21 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
+use App\Mail\ProjectCreated;
 use App\Project;
+
 
 class ProjectsController extends Controller
 {
+    public function __construct() {
+        $this->middleware('auth');
+    }
+
     public function index()
     {
-        $projects = \App\Project::all();
-
-        return view('projects.index', compact('projects'));
+        return view('projects.index', [
+            'projects' => auth()->user()->projects
+        ]);
     }
 
     public function create()
@@ -21,23 +26,35 @@ class ProjectsController extends Controller
 
     public function show(Project $project)
     {
+        // abort_if($project->owner_id !== auth()->id(), 403);
+        // abort_unless($project->owner_id == auth()->id(), 403);
+        // abort_unless (\Gate::allows('update', $project), 403);
+        // abort_if (\Gate::denies('update', $project), 403);
+        $this->authorize('update', $project);
+
         return view('projects.show', compact('project'));
     }
 
     public function edit(Project $project)
     {
+        $this->authorize('update', $project);
+        
         return view('projects.edit', compact('project'));
     }
 
     public function update(Project $project)
     {
-        $project->update(request(['title', 'description']));
+        $this->authorize('update', $project);
+
+        $project->update($this->validateProject());
         
         return redirect('/projects');
     }
 
     public function destroy(Project $project)
     {
+        $this->authorize('update', $project);
+        
         $project->delete();
 
         return redirect('/projects');
@@ -45,13 +62,22 @@ class ProjectsController extends Controller
 
     public function store()
     {
-        Project::create(
-            request()->validate([
-                'title' => 'required',
-                'description' => 'required'
-            ])
-        );
+        $attributes = $this->validateProject();
+
+        $attributes['owner_id'] = auth()->id();
+        
+        $project = Project::create($attributes);
+
+        event(new ProjectCreated($project));
 
         return redirect('/projects');
+    }
+
+    protected function validateProject()
+    {
+        return request()->validate([
+            'title' => ['required', 'min:3'],
+            'description' => ['required', 'min:3']
+        ]);
     }
 }
